+++ changelog.md
@@ -1,6 +1,12 @@
 # PHPMailer Change Log
 
+## Version 6.1.1 (September 27th 2019)
+* Fix misordered version tag
+
+## Version 6.1.0 (September 27th 2019)
+* Multiple bug fixes for folding of long header lines, thanks to @caugner
 * Add support for [RFC2387 child element content-type hint](https://tools.ietf.org/html/rfc2387#section-3.1) in `multipart/related` structures.
+* Support for Ical event methods other than `REQUEST`, thanks to @puhr-mde
 * Change header folding and param separation to use spaces instead of tabs
 * Use ; to separate multiple MIME header params
 * Add support for RFC3461 DSN messages
@@ -10,10 +16,15 @@
 * Updated Brazilian Portuguese translation
 * Throw exceptions on invalid encoding values
 * Add Afrikaans translation, thanks to @Donno191
+* Updated Farsi/Persian translation
 * Add PHP 7.4 to test config
 * Remove some ambiguity about setting XMailer property
 * Improve error checking in mailing list example
 * Drop PHP 5.5 from CI config as it's no longer supported by Travis-CI
+* Fix S/MIME signing
+* Add constants for encryption type
+* More consistent use of constants for encryption, charset, encoding
+* Add PHPMailer logo images
 
 ## Version 6.0.7 (February 1st 2019)
 * Include RedHat GPL Cooperation Commitment - see the `COMMITMENT` file for details.